[SPARK-40640][CORE] SparkHadoopUtil to set origin of hadoop/hive config options
### What changes were proposed in this pull request?

The options passed from spark conf, hive-site.xml, AWS env vars now all record this in their source attribute of the entries.

The Configuration Writable methods do not propagate this, so it is not as useful cluster-wide than it could be. It does help with some of the basic troubleshooting.

### Why are the changes needed?

Helps when troubleshooting where options make their way down. These can be examined
and logged later.

For example, my cloudstore diagnosticss JAR can do this in its storediag command
and in an s3a AWS credential provider. I may add some of that logging
at debug to the ASF hadoop implementations.

https://github.com/steveloughran/cloudstore

### Does this PR introduce _any_ user-facing change?

Not *really*. It's a very low level diagnostics feature in the Hadoop configuration classes.

### How was this patch tested?

New tests added; existing tests enhanced.

Closes #38084 from steveloughran/SPARK-40640-spark-conf-propagation.

Lead-authored-by: Steve Loughran <stevel@cloudera.com>
Co-authored-by: Steve Loughran <stevel@apache.org>
Co-authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>