[SPARK-40737][CONNECT] Add basic support for DataFrameWriter
### What changes were proposed in this pull request?

This change adds basic support for writes through the Spark Connect API. In todays implementation of the write API from a DataFrame perspective the interface of the DataFrameWriter is as declarative as possible today.

The write support is implemented as a `Command` and does not return anything.

### Why are the changes needed?

Write support through Spark Connect.

### Does this PR introduce _any_ user-facing change?
Experimental API

### How was this patch tested?

Added new unit tests for the behavior.

Closes #38192 from grundprinzip/spark-40737.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>