[SPARK-40957] Add in memory cache in HDFSMetadataLog
### What changes were proposed in this pull request?

Every time entries in offset log or commit log needs to be access, we read from disk which is slow.  Can a cache of recent entries to speed up reads.

There is already an existing implementation of a caching mechanism in OffsetSeqLog.  Lets replace it with an implementation in HDFSMetadataLog (parent class) so that we can support reading from in memory cache for both offset log and commit log.

### Why are the changes needed?

Improve read speeds for entries in offset log and commit log

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing unit tests should suffice

Closes #38430 from jerrypeng/SPARK-40957.

Authored-by: Jerry Peng <jerry.peng@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>