[SPARK-34079][SQL][FOLLOWUP] Improve the readability and simplify the code for MergeScalarSubqueries
### What changes were proposed in this pull request?
Recently, I read the `MergeScalarSubqueries` because it is a feature used for improve performance.
I fount the parameters of ScalarSubqueryReference is hard to understand, so I want add some comments on it.

Additionally, the private method `supportedAggregateMerge` of `MergeScalarSubqueries` looks redundant, this PR wants simplify the code.

### Why are the changes needed?
Improve the readability and simplify the code for `MergeScalarSubqueries`.

### Does this PR introduce _any_ user-facing change?
'No'.
Just improve the readability and simplify the code for `MergeScalarSubqueries`.

### How was this patch tested?
Exists tests.

Closes #38461 from beliefer/SPARK-34079_followup.

Authored-by: Jiaan Geng <beliefer@163.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>