[SPARK-40815][SQL][FOLLOW-UP] Disable DelegateSymlinkTextInputFormat tests for JDK 9+
### What changes were proposed in this pull request?

This PR is a follow-up for https://github.com/apache/spark/pull/38277.

This change is required due to test failures in JDK 11 and JDK 17. The patch disables the unit test for JDK 9+.
This is a temporary measure while I am debugging and working on the fix for higher versions of JDK.

### Why are the changes needed?

Fixes the test failure in JDK 11.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

N/A.

Closes #38504 from sadikovi/fix-symlink-test.

Authored-by: Ivan Sadikov <ivan.sadikov@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>