[SPARK-41019][SQL] Provide a query context to `failAnalysis()`
### What changes were proposed in this pull request?
In the PR, I propose to invoke `AnalysisErrorAt.failAnalysis()` instead of `CheckAnalysis.failAnalysis()` because the first one captures the query context and passes it to `AnalysisException`.

### Why are the changes needed?
To provide additional info as a query context to users. This should improve user experience with Spark SQL.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
By running the modified test suites:
```
$ PYSPARK_PYTHON=python3 build/sbt "sql/testOnly org.apache.spark.sql.SQLQueryTestSuite"
```

Closes #38514 from MaxGekk/provide-context-failAnalysis.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>