[SPARK-39883][SQL][TESTS] Add DataFrame function parity check
### What changes were proposed in this pull request?

This PR adds a test that compares the available list of DataFrame functions in org.apache.spark.sql.functions with the SQL function registry. This attempts to verify that the DataFrame functions are a subset of the functions in the SQL function registry (subject to exclusions and expectations). ~It also produces a list of the differences between the two.~(removed per feedback)

### Why are the changes needed?

Currently there is no easy way to verify what the difference is between the two API's

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

This PR is testing only

Closes #37303 from aray/scala-function-parity-test.

Authored-by: Andrew Ray <ray.andrew@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>