[SPARK-41014][PYTHON][DOC] Improve documentation and typing of groupby and cogroup applyInPandas
### What changes were proposed in this pull request?
Documentation of method `applyInPandas` for grouby and cogroup should mention in the main description that there are two allowed signatures for the provided function. The Examples state that this is possible, and the parameters sections states that for cogroup.

Also type information for `PandasCogroupedMapFunction` should mention the three-argument callable alternative.

### Why are the changes needed?
Better documenting these alternative to find this piece of information quicker.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Untested

Closes #38509 from EnricoMi/branch-pyspark-group-and-cogroup-doc-and-typing.

Authored-by: Enrico Minack <github@enrico.minack.dev>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>